requirements.txt
streamlit 
langchain-openai 
langchain-groq 
langchain-google-genai 
langchain-core


langchain-community
faiss-cpu
sentence-transformers
python-dotenv
google-api-python-client
unstructured